Add getConnections method to config
