+++ server/test/models/connections.js
@@ -1,8 +1,8 @@
 const assert = require('assert');
+const { config } = require('../utils');
 const { getDb } = require('../../lib/db');
-const { getConnectionsFromConfig } = require('../../lib/connectionsFromConfig');
 
-describe('getConnectionsFromConfig', function() {
+describe('config.getConnections', function() {
   let models;
 
   before(async function() {
@@ -11,12 +11,12 @@ describe('getConnectionsFromConfig', function() {
   });
 
   it('handles empty object', function() {
-    const cs = getConnectionsFromConfig({});
+    const cs = config.getConnections({});
     assert(Array.isArray(cs));
   });
 
   it('skips partial connections', function() {
-    const cs = getConnectionsFromConfig({
+    const cs = config.getConnections({
       SQLPAD_CONNECTIONS__abc__driver: 'postgres'
     });
     assert(Array.isArray(cs));
@@ -24,7 +24,7 @@ describe('getConnectionsFromConfig', function() {
   });
 
   it('parses connection properly', function() {
-    const cs = getConnectionsFromConfig({
+    const cs = config.getConnections({
       SQLPAD_CONNECTIONS__abc__driver: 'postgres',
       SQLPAD_CONNECTIONS__abc__name: 'env-postgres',
       SQLPAD_CONNECTIONS__abc__host: 'localhost',
